+++ src/main/groovy/com/ullink/AssemblyInfoVersionPatcher.groovy
@@ -44,6 +44,9 @@ class AssemblyInfoVersionPatcher extends ConventionTask {
 
     @Input
     def informationalVersion
+    
+    @Input
+    def assemblyDescription = ''
 
     @Input
     def charset = "UTF-8"
@@ -61,12 +64,19 @@ class AssemblyInfoVersionPatcher extends ConventionTask {
             replace(it, 'AssemblyVersion', getVersion())
             replace(it, 'AssemblyFileVersion', getFileVersion())
             replace(it, 'AssemblyInformationalVersion', getInformationalVersion())
+            replace(it, 'AssemblyDescription', getAssemblyDescription())
         }
     }
 
     void replace(def file, def name, def value) {
+        //only change the assembly values if they specified here (not blank or null)
+        //if the parameters are blank, then keep whatever is already in the assemblyinfo file.
+        if (value == null || value.isEmpty()) return
+    
         if (FilenameUtils.getExtension(file.name) == 'fs')
             project.ant.replaceregexp(file: file, match: /^\[<assembly: $name\s*\(".*"\)\s*>\]$/, replace: "[<assembly: ${name}(\"${value}\")>]", byline: true, encoding: charset)
+        else if (FilenameUtils.getExtension(file.name) == 'vb')
+            project.ant.replaceregexp(file: file, match: /^<Assembly: $name\s*\(".*"\)\s*>$/, replace: "<Assembly: ${name}(\"${value}\")>", byline: true, encoding: charset)
         else
             project.ant.replaceregexp(file: file, match: /^\[assembly: $name\s*\(".*"\)\s*\]$/, replace: "[assembly: ${name}(\"${value}\")]", byline: true, encoding: charset)
     }